+++ BraintreeCore/src/main/java/com/braintreepayments/api/core/AnalyticsEventParams.kt
@@ -23,12 +23,12 @@ import androidx.annotation.RestrictTo
  */
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 data class AnalyticsEventParams @JvmOverloads constructor(
-    var payPalContextId: String? = null,
-    var linkType: String? = null,
-    var isVaultRequest: Boolean = false,
-    var startTime: Long? = null,
-    var endTime: Long? = null,
-    var endpoint: String? = null,
+    val payPalContextId: String? = null,
+    val linkType: String? = null,
+    val isVaultRequest: Boolean = false,
+    val startTime: Long? = null,
+    val endTime: Long? = null,
+    val endpoint: String? = null,
     val experiment: String? = null,
     val appSwitchUrl: String? = null,
     val shopperSessionId: String? = null,