+++ PayPal/src/test/java/com/braintreepayments/api/paypal/PayPalClientUnitTest.java
@@ -286,8 +286,11 @@ public class PayPalClientUnitTest {
                 "for more information.",
             ((PayPalPaymentAuthRequest.Failure) request).getError().getMessage());
 
-        AnalyticsEventParams params = new AnalyticsEventParams();
-        params.setVaultRequest(false);
+        AnalyticsEventParams params = new AnalyticsEventParams(
+            null,
+            null,
+            false
+        );
         verify(braintreeClient).sendAnalyticsEvent(PayPalAnalytics.TOKENIZATION_FAILED, params);
     }
 
@@ -311,8 +314,11 @@ public class PayPalClientUnitTest {
         assertTrue(request instanceof PayPalPaymentAuthRequest.Failure);
         assertEquals(authError, ((PayPalPaymentAuthRequest.Failure) request).getError());
 
-        AnalyticsEventParams params = new AnalyticsEventParams();
-        params.setVaultRequest(false);
+        AnalyticsEventParams params = new AnalyticsEventParams(
+            null,
+            null,
+            false
+        );
         verify(braintreeClient).sendAnalyticsEvent(PayPalAnalytics.TOKENIZATION_FAILED, params);
     }
 
@@ -336,8 +342,11 @@ public class PayPalClientUnitTest {
         assertTrue(request instanceof PayPalPaymentAuthRequest.Failure);
         assertEquals(authError, ((PayPalPaymentAuthRequest.Failure) request).getError());
 
-        AnalyticsEventParams params = new AnalyticsEventParams();
-        params.setVaultRequest(true);
+        AnalyticsEventParams params = new AnalyticsEventParams(
+            null,
+            null,
+            true
+        );
         verify(braintreeClient).sendAnalyticsEvent(PayPalAnalytics.TOKENIZATION_FAILED, params);
     }
 
@@ -416,9 +425,11 @@ public class PayPalClientUnitTest {
         assertFalse(browserSwitchOptions.isLaunchAsNewTask());
 
 
-        AnalyticsEventParams params = new AnalyticsEventParams();
-        params.setVaultRequest(true);
-        params.setLinkType("universal");
+        AnalyticsEventParams params = new AnalyticsEventParams(
+            null,
+            "universal",
+            true
+        );
         verify(braintreeClient).sendAnalyticsEvent(PayPalAnalytics.APP_SWITCH_STARTED, params);
     }
 
@@ -543,8 +554,11 @@ public class PayPalClientUnitTest {
         PayPalResult result = captor.getValue();
         assertTrue(result instanceof PayPalResult.Cancel);
 
-        AnalyticsEventParams params = new AnalyticsEventParams();
-        params.setVaultRequest(false);
+        AnalyticsEventParams params = new AnalyticsEventParams(
+            null,
+            null,
+            false
+        );
         verify(braintreeClient).sendAnalyticsEvent(PayPalAnalytics.BROWSER_LOGIN_CANCELED, params);
     }
 
@@ -584,9 +598,11 @@ public class PayPalClientUnitTest {
         assertTrue(result instanceof PayPalResult.Success);
         assertEquals(payPalAccountNonce, ((PayPalResult.Success) result).getNonce());
 
-        AnalyticsEventParams params = new AnalyticsEventParams();
-        params.setPayPalContextId("EC-HERMES-SANDBOX-EC-TOKEN");
-        params.setVaultRequest(false);
+        AnalyticsEventParams params = new AnalyticsEventParams(
+            "EC-HERMES-SANDBOX-EC-TOKEN",
+            null,
+            false
+        );
         verify(braintreeClient).sendAnalyticsEvent(PayPalAnalytics.TOKENIZATION_SUCCEEDED, params);
     }
 
@@ -626,8 +642,9 @@ public class PayPalClientUnitTest {
         assertTrue(result instanceof PayPalResult.Success);
         assertEquals(payPalAccountNonce, ((PayPalResult.Success) result).getNonce());
 
-        AnalyticsEventParams params = new AnalyticsEventParams();
-        params.setPayPalContextId("EC-HERMES-SANDBOX-EC-TOKEN");
+        AnalyticsEventParams params = new AnalyticsEventParams(
+            "EC-HERMES-SANDBOX-EC-TOKEN"
+        );
         verify(braintreeClient).sendAnalyticsEvent(PayPalAnalytics.TOKENIZATION_SUCCEEDED, params);
         AnalyticsEventParams appSwitchParams = new AnalyticsEventParams(
                 "EC-HERMES-SANDBOX-EC-TOKEN",
@@ -667,8 +684,9 @@ public class PayPalClientUnitTest {
 
         sut.tokenize(payPalPaymentAuthResult, payPalTokenizeCallback);
 
-        AnalyticsEventParams params = new AnalyticsEventParams();
-        params.setPayPalContextId("SOME-BA");
+        AnalyticsEventParams params = new AnalyticsEventParams(
+            "SOME-BA"
+        );
         verify(braintreeClient).sendAnalyticsEvent(PayPalAnalytics.TOKENIZATION_FAILED, params);
         AnalyticsEventParams appSwitchParams = new AnalyticsEventParams(
                 "SOME-BA",