+++ CHANGELOG.md
@@ -7,8 +7,11 @@
     * Add `shopperSessionId` parameter to `ShopperInsightsClient`
 * BraintreePayPal
     * Add `shopperSessionId` to `PayPalCheckoutRequest` and `PayPalVaultRequest`
+    * Fix bug to ensure that `PayPalVaultRequest.userAuthenticationEmail` is not sent as an empty string
     * Replace `sendPayPalPresentedEvent()` and `sendVenmoPresentedEvent()` with `sendPresentedEvent()`
     * Replace `sendPayPalSelectedEvent()` and `sendVenmoSelectedEvent()` with `sendSelectedEvent()`
+* ThreeDSecure
+    * Return error if no `dfReferenceId` is returned in the 3D Secure flow
 
 ## 5.3.0 (2024-12-11)